services/identity-service/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
requests
uuid
azure-ai-vision-face